+++ packages/expo-router/src/native-tabs/__tests__/NativeTabsView.test.ios.tsx
@@ -30,7 +30,7 @@ it.each([
   { value: false, expected: 'tabBar' },
 ] as {
   value: NativeTabsProps['sidebarAdaptable'];
-  expected: TabsHostProps['tabBarControllerMode'];
+  expected: NonNullable<TabsHostProps['ios']>['tabBarControllerMode'];
 }[])('when sidebarAdaptable is $value, then ', ({ value, expected }) => {
   renderRouter({
     _layout: () => (
@@ -43,7 +43,7 @@ it.each([
 
   expect(screen.getByTestId('index')).toBeVisible();
   expect(TabsHost).toHaveBeenCalledTimes(1);
-  expect(TabsHost.mock.calls[0]![0].tabBarControllerMode).toBe(expected);
+  expect(TabsHost.mock.calls[0][0].ios?.tabBarControllerMode).toBe(expected);
 });
 
 it('uses shadowColor when it is passed to NativeTabs', () => {
@@ -58,8 +58,10 @@ it('uses shadowColor when it is passed to NativeTabs', () => {
 
   expect(screen.getByTestId('index')).toBeVisible();
   expect(TabsScreen).toHaveBeenCalledTimes(1);
-  expect(TabsScreen.mock.calls[0]![0].standardAppearance!.tabBarShadowColor).toBe('red');
-  expect(TabsScreen.mock.calls[0]![0].scrollEdgeAppearance!.tabBarShadowColor).toBe('transparent');
+  expect(TabsScreen.mock.calls[0][0].ios?.standardAppearance!.tabBarShadowColor).toBe('red');
+  expect(TabsScreen.mock.calls[0][0].ios?.scrollEdgeAppearance!.tabBarShadowColor).toBe(
+    'transparent'
+  );
 });
 
 it('uses shadowColor when it is passed to NativeTabs in both standardAppearance and scrollEdgeAppearance when disableTransparentOnScrollEdge is true', () => {
@@ -74,6 +76,6 @@ it('uses shadowColor when it is passed to NativeTabs in both standardAppearance 
 
   expect(screen.getByTestId('index')).toBeVisible();
   expect(TabsScreen).toHaveBeenCalledTimes(1);
-  expect(TabsScreen.mock.calls[0]![0].standardAppearance!.tabBarShadowColor).toBe('red');
-  expect(TabsScreen.mock.calls[0]![0].scrollEdgeAppearance!.tabBarShadowColor).toBe('red');
+  expect(TabsScreen.mock.calls[0][0].ios?.standardAppearance!.tabBarShadowColor).toBe('red');
+  expect(TabsScreen.mock.calls[0][0].ios?.scrollEdgeAppearance!.tabBarShadowColor).toBe('red');
 });